+++ src/distribution/Pie.ts
@@ -14,7 +14,7 @@ import {ICatHistogram} from 'phovea_core/src/math';
 import {toSelectOperation} from 'phovea_core/src/idtype';
 import {vec2, polygon} from 'phovea_core/src/geom';
 import bindTooltip from 'phovea_d3/src/tooltip';
-import {IDistributionOptions, ITotalHeight, resolveHistMax} from './utils';
+import {IDistributionOptions, ITotalHeight, resolveHistMax} from './internal';
 
 
 function toPolygon(start: number, end: number, radius: number) {
