+++ src/heatmap/HeatMapCanvasRenderer.ts
@@ -5,7 +5,7 @@
 
 import * as d3 from 'd3';
 import {IMatrix} from 'phovea_core/src/matrix';
-import {IScale} from './utils';
+import {IScale} from './internal';
 import {IHeatMapRenderer, ESelectOption} from './IHeatMapRenderer';
 import AHeatMapCanvasRenderer from './AHeatMapCanvasRenderer';
