+++ src/heatmap/internal.ts
